Naprawa błędu braku odrzucania nowych rekordów w aktualizacji

+++ frontend/app/src/main/java/com/example/frontend/ui/service/ApiService.kt
@@ -112,7 +112,7 @@ interface ApiService {
 
     @PATCH("ingredients/{id}")
     suspend fun updateIngredient(
-        @Path("id") id: Int,
+        @Path("id") id: Float,
         @Body body: UpdateIngredientRequest
     ): Response<Ingredient>
